Update to mytile.py
This is a cosmetic change on the graph dials to remove trailing 0s on float unit measurements that do not actually have hundredth or tenth precision.  
For example, the Battery Voltage registers only publish 1/10 voltage, and Hz only Integer.  On the default dial, those are represented with 2 digit precision after the period.  
  
This simply changes the two digit float (%2f) to a float with trailing 0s removed (%g), handles different measurement precision by arbitrarily omitting trailing 0s and makes the dials simpler to read.